[LCR-185] refactor: item-service 와 분리 후 디렉토리 구조 변경
