+++ popi-popup-service/src/main/java/com/lgcns/dto/response/PopupInfoResponse.java
@@ -1,4 +1,4 @@
-package com.lgcns.dto.popup.response;
+package com.lgcns.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
