+++ popi-popup-service/src/main/java/com/lgcns/service/PopupService.java
@@ -1,6 +1,6 @@
-package com.lgcns.service.popup;
+package com.lgcns.service;
 
-import com.lgcns.dto.popup.response.PopupInfoResponse;
+import com.lgcns.dto.response.PopupInfoResponse;
 import com.lgcns.response.SliceResponse;
 
 public interface PopupService {
