+++ popi-popup-service/src/test/java/com/lgcns/service/PopupServiceTest.java
@@ -6,9 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.lgcns.WireMockIntegrationTest;
-import com.lgcns.dto.popup.response.PopupInfoResponse;
+import com.lgcns.dto.response.PopupInfoResponse;
 import com.lgcns.response.SliceResponse;
-import com.lgcns.service.popup.PopupService;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Assertions;
